# backend/schema/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "ShieldPM API",
		"version": "2.x.x",
		"description": "This is the official API documentation for ShieldPM.\n\nMost endpoints require authentication via Bearer Token (JWT). You can generate a token by logging in via the `POST /tokens` endpoint.\n\nFor more information, visit the [ShieldPM GitHub Repository](https://github.com/shedowe19/ShieldPM)."
	},
	"servers": [
		{
			"url": "http://127.0.0.1:81/api"
		}
	],
	"components": {
		"securitySchemes": {
			"$ref": "./components/security-schemes.json"
		}
	},
...
	"paths": {
...
		"/ai/config": {
			"get": {
				"tags": [
					"AI"
				],
				"summary": "Get AI Configuration",
				"responses": {
					"200": {
						"description": "Configuration object",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				]
			},
			"put": {
				"$ref": "./paths/ai/config.json"
			}
		},
...
	}
}